requirements.txt
# Core LangChain Orchestration
langchain>=1.0.0,<2.0.0                
langchain-huggingface>=0.0.17,<1.0.0   
langchain-community>=0.1.27,<1.0.0     
langchain-google-genai>=0.0.15,<1.0.0  

# Vectorstore
chromadb>=0.3.30,<1.0.0

# Embedding Models
sentence-transformers>=2.2.2,<3.0.0    
torch>=2.0.0,<3.0.0                    

# Local LLM Wrappers
llama-cpp-python>=0.1.81,<1.0.0       

# Core Utilities
pandas>=2.1.0,<3.0.0                 
python-dotenv>=1.0.0,<2.0.0          
tqdm>=4.66.1,<5.0.0                    
typing-extensions>=4.8.0,<5.0.0      